+++ src/containers/Users/index.jsx
@@ -1,34 +1,22 @@
-import React, { useCallback, useEffect, useState } from "react";
-import { Button, Menu, Select, Table } from "antd";
+import React, { useEffect } from "react";
+import { Table } from "antd";
 import { columns } from "./columns";
-import { useNavigate } from "react-router-dom";
-import { fetchUsersApi } from "../../api";
-import { useDispatch, useSelector } from "react-redux";
-import { fetchUsers, fetchUsersFailed, fetchUsersSuccess } from "./actions";
+import { useNavigate, useParams } from "react-router-dom";
+import { fetchUsers } from "./actions";
+import withDynamicColumnsHOC from "../../components/Table/withDynamicColumnsHOC";
+import useInitialFetch from "../../hooks/useInitialFetch";
 
-const { Option } = Select;
+const DynamicTable = withDynamicColumnsHOC(Table);
 
 export default function Users() {
-  const [tableColumns, setTableColumns] = useState(columns);
+  const { id } = useParams();
 
-  const dispatch = useDispatch();
-  const state = useSelector((store) => store.users);
-  const lastLoadedUser = useSelector((store) => store.user.user);
-  const { data, isLoading, error } = state;
+  const state = useInitialFetch(fetchUsers, {
+    payload: id,
+    stateSelector: "users",
+  });
 
-  const handleColumnChange = useCallback(
-    (selection) => {
-      const newColumns = columns.filter(({ title }) =>
-        selection.includes(title.toLowerCase())
-      );
-      setTableColumns(newColumns);
-    },
-    [tableColumns]
-  );
-
-  useEffect(() => {
-    dispatch(fetchUsers());
-  }, []);
+  const { data } = state;
 
   const navigate = useNavigate();
   const onRowClick = (row) => {
@@ -38,22 +26,8 @@ export default function Users() {
 
   return (
     <>
-      <div>Last selected user name: {lastLoadedUser?.name}</div>
-      <Select
-        mode="multiple"
-        style={{ width: "30%" }}
-        defaultValue={tableColumns.map(({ title }) => title.toLowerCase())}
-        placeholder="Please select columns"
-        onChange={handleColumnChange}
-      >
-        {columns.map((column) => (
-          <Option key={column.key} disabled={column.required}>
-            {column.title}
-          </Option>
-        ))}
-      </Select>
-      <Table
-        columns={tableColumns}
+      <DynamicTable
+        columns={columns}
         dataSource={data}
         onRow={(record) => ({
           onDoubleClick: () => onRowClick(record),
